fix TLS pointer size to be width independent

+++ examples/udp-bindings/src/main.rs
@@ -98,7 +98,7 @@ async fn run_wasm() -> wasmtime::Result<()> {
 // Same as https://github.com/bytecodealliance/wasmtime/blob/main/examples/min-platform/embedding/wasmtime-platform.c
 // I have no idea whether this is safe or not.
 // https://github.com/bytecodealliance/wasmtime/blob/aec935f2e746d71934c8a131be15bbbb4392138c/crates/wasmtime/src/runtime/vm/traphandlers.rs#L888
-static mut TLS_PTR: u32 = 0;
+static mut TLS_PTR: usize = 0;
 #[unsafe(no_mangle)]
 extern "C" fn wasmtime_tls_get() -> *mut u8 {
     unsafe { TLS_PTR as *mut u8 }
@@ -106,5 +106,5 @@ extern "C" fn wasmtime_tls_get() -> *mut u8 {
 
 #[unsafe(no_mangle)]
 extern "C" fn wasmtime_tls_set(val: *const u8) {
-    unsafe { TLS_PTR = val as u32 };
+    unsafe { TLS_PTR = val as usize };
 }
